Recategorize test
TestInstallFromTag_ManageFromClaim hits docker and should be
an integration test
