+++ pkg/porter/lifecycle_test.go
@@ -185,32 +185,3 @@ func TestManifestIgnoredWithTag(t *testing.T) {
 		require.NoError(t, err, "Validate failed")
 	})
 }
-
-func TestInstallFromTag_ManageFromClaim(t *testing.T) {
-	p := NewTestPorter(t)
-	cacheDir, _ := p.Cache.GetCacheDir()
-	p.TestConfig.TestContext.AddTestDirectory("testdata/cache", cacheDir)
-
-	installOpts := NewInstallOptions()
-	installOpts.Name = "hello"
-	installOpts.Tag = "getporter/porter-hello:v0.1.1"
-	err := installOpts.Validate(nil, p.Porter)
-	require.NoError(t, err, "InstallOptions.Validate failed")
-
-	err = p.InstallBundle(installOpts)
-	require.NoError(t, err, "InstallBundle failed")
-
-	upgradeOpts := NewUpgradeOptions()
-	upgradeOpts.Name = installOpts.Name
-	err = upgradeOpts.Validate(nil, p.Porter)
-
-	err = p.UpgradeBundle(upgradeOpts)
-	require.NoError(t, err, "UpgradeBundle failed")
-
-	uninstallOpts := NewUninstallOptions()
-	uninstallOpts.Name = installOpts.Name
-	err = uninstallOpts.Validate(nil, p.Porter)
-
-	err = p.UninstallBundle(uninstallOpts)
-	require.NoError(t, err, "UninstallBundle failed")
-}
